12/24 19:31 By JAY

+++ lib/pages/3_personalInformationPage/personalInfoData.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class personalInfoData {
   String univ;
   String major;
-  int gpa;
+  double gpa;
   int? age;
   String? gender;
